Refactor: Reorganización de estructura MVC estricta en modelo, vista y controlador
