+++ yugi_oh_JAVA/src/App.java
@@ -1,3 +1,5 @@
+
+
 import javax.swing.SwingUtilities;
 import controlador.ControladorJuego;
 import vista.MenuInicial;
