+++ yugi_oh_JAVA/src/controlador/Comando.java
@@ -1,4 +1,4 @@
-package patrones;
+package controlador;
 
 public interface Comando {
     void ejecutar();
