+++ yugi_oh_JAVA/src/modelo/CampoLlenoException.java
@@ -1,4 +1,4 @@
-package excepciones;
+package modelo;
 
 
 public class CampoLlenoException extends Exception {
