+++ yugi_oh_JAVA/src/modelo/CargadorCartas.java
@@ -1,4 +1,4 @@
-package persistencia;
+package modelo;
 
 import modelo.Carta;
