+++ yugi_oh_JAVA/src/modelo/GestorMementos.java
@@ -1,4 +1,4 @@
-package patrones;
+package modelo;
 
 import java.util.Stack;
