+++ yugi_oh_JAVA/src/modelo/Jugador.java
@@ -3,7 +3,7 @@ package modelo;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.Stack;
-import excepciones.MazoVacioException;
+import modelo.MazoVacioException;
 
 public class Jugador {
