+++ yugi_oh_JAVA/src/modelo/MazoVacioException.java
@@ -1,4 +1,4 @@
-package excepciones;
+package modelo;
 
 
 public class MazoVacioException extends Exception {
