+++ yugi_oh_JAVA/src/modelo/Observable.java
@@ -1,5 +1,6 @@
-package patrones;
+package modelo;
 
+import vista.Observador;
 import java.util.ArrayList;
 import java.util.List;
