+++ yugi_oh_JAVA/src/modelo/TurnoInvalidoException.java
@@ -1,4 +1,4 @@
-package excepciones;
+package modelo;
 
 
 public class TurnoInvalidoException extends Exception {
