+++ yugi_oh_JAVA/src/vista/Observador.java
@@ -1,4 +1,4 @@
-package patrones;
+package vista;
 
 public interface Observador {
     void actualizar();
